pyproject.toml
[project]
name = "graphbrain"
version = "0.7.0-dev"
description = "Semantic Hypergraph Tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Telmo Menezes et al.", email = "telmo@telmomenezes.net" },
]
keywords = [
    "NLP",
    "AI",
    "Knowledge Representation",
    "Knowledge Systems",
    "Natural Language Understanding",
    "Text Analysis",
    "Cognition",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "asciitree",
    "ipython",
    "mwparserfromhell",
    "networkx",
    # fixed for now because of spacy
    "numpy", # ==1.26.4',
    "progressbar2",
    "scikit-learn",
    "spacy",
    "spacy-transformers",
    "termcolor",
    "thinc",
    "torch",
    "trafilatura",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "ruff>=0.2.2",
    "pre-commit>=3.6.2",
    "types-passlib>=1.7.7.20240106",
    "coverage>=7.4.3",
    "Sphinx",
    "sphinx_rtd_theme",
]

[project.urls]
Homepage = "https://graphbrain.net"

[project.scripts]
graphbrain = "graphbrain.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphbrain"]

[tool.mypy]
strict = true

[tool.ruff]
target-version = "py310"
